Add auto restart feature
Restart Layer brush feature and create a new CaptureService if display settings is changed.

This commit also fix Exceptions when capture area goes beyond layer display resolution boundaries.

I created a feature as it appear to be the cleaner way to do this but not 100% sure if i am correct.