Fix empty ripple center under certain configurations
SkiaSharp shapes doesn't support inner stroke. It causes a weird empty circle if the stroke width is greater than the double of the size of a shape. This is noticeable only under certain conditions like low speed and wider ripples. This operation will produce perfect ripples from start to finish.